+++ types/core/src/index.d.ts
@@ -15,6 +15,7 @@ export * from "./invert.js";
 export * from "./placement.js";
 export * from "./area.js";
 export * from "./canvasitem.js";
+export * from "./googledoc.js";
 export * from "./titleroom.js";
 export * from "./whatsnew.js";
 export * from "./protocol.js";